better usb reset function
I discovered a hack:  "gphoto2 -l", when run, unblocks the canon camera so it will operate again on the back display; the old method was to disable power to the usb port. This works much better and runs faster.

Additionally, added some packages to support the port to the rpi zero 2w.